Перевантажено оператори +, -, * і оператор неявного приведення double до Polynomial для класу Polynomial

+++ Задача 6.1/PolynomialPart.cs
@@ -2,10 +2,10 @@
 {
     public class PolynomialPart
     {
-        public int RealPart { get; set; }
+        public double RealPart { get; set; }
         public int PolynomRank { get; set; }
 
-        public PolynomialPart(int realPart, int polynomRank)
+        public PolynomialPart(double realPart, int polynomRank)
         {
             RealPart = realPart;
             PolynomRank = polynomRank;
